pyproject.toml
[tool.poetry]
name = "toktik-thumbnailer"
version = "0.1.0"
description = ""
authors = ["thanawinboon <pokepat27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.28.70"
celery = "^5.3.4"
python-dotenv = "^1.0.0"
opencv-python = "^4.8.1.78"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
